Move the addEditHandler function to take advantage of its use after other item is added to the list
