+++ src/add_remove.js
@@ -1,6 +1,23 @@
 import { getFromLocalStorage, setToLocalStorage } from "./storage";
 import dragAndDrop from './drag_drop.js';
 
+document.querySelector('.todo-new > input').addEventListener('keypress', (e) => {
+  if (e.key === 'Enter') { 
+    addTodo(e.target.value);
+    e.target.value = "";
+  }
+});
+
+export function addEditHandlers() {
+  const todoList = document.getElementsByClassName('todo-item');
+  for (let i = 0; i < todoList.length; i += 1) {
+    let labelElem = todoList[i].children[0].children[1];
+    labelElem.addEventListener('input', () => {
+      editTodo(labelElem);
+    });
+  }
+}
+
 function appendToDOM(todo) {
   document.getElementById('todo-list').insertAdjacentHTML('beforeend', `
     <div class="todo-item" draggable="true">
@@ -25,12 +42,19 @@ export function addTodo(description) {
 
   let currentTodoList = getFromLocalStorage();
   let todoLength = currentTodoList.length;
-  newTodo.index = todoLength + 1;
+  if (todoLength === 0) {
+    newTodo.index = 0;
+  } else {
+    newTodo.index = todoLength;
+  }
+  
   
   currentTodoList.push(newTodo);
   setToLocalStorage(currentTodoList);
   appendToDOM(newTodo);
   dragAndDrop();
+
+  addEditHandlers();
 }
 
 export function editTodo(item) {
